feat: Add count total on REST endpoints
