+++ src/services/api/utils.ts
@@ -1,20 +1,5 @@
 import { ValidationError } from '../../exceptions/validation';
-
-/**
- * GET query definition
- */
-export interface Query {
-  name?: string;
-  limit?: number;
-  offset?: number;
-}
-
-/**
- * GET param definition
- */
-export interface Params {
-  id: number;
-}
+import { Query, Params } from './const';
 
 /**
  * Validate user inputs and transform them into Query
@@ -78,25 +63,3 @@ export function validateParam(param: any): Params {
     id: convertedValue,
   };
 }
-
-/**
- * GET API response format
- */
-export interface OJKData<T> {
-  data: T | T[] | null;
-  version: string;
-}
-
-export enum HTTPCodes {
-  // HTTP status code when input is invalid
-  INVALID_PARAMS = 400,
-  // HTTP status code when data is not found
-  NOT_FOUND = 404,
-  // HTTP status code when JSON files aren't created or unexpected error occurs
-  SERVER_ERROR = 500,
-  // HTTP status code when the request method isn't GET
-  METHOD_NOT_ALLOWED = 405,
-  // HTTP status code when request has been successfully handled
-  SUCCESS = 200,
-};
-
